finish offload controller, split up all new files
