+++ lance/src/robot/controllers/auto_mining_controller.hpp
@@ -92,85 +92,3 @@ protected:
     TraversalController& traversal_controller;
     MiningController mining_controller;
 };
-
-
-// --- Implementation ----------------------------------------------------------
-
-AutoMiningController::AutoMiningController(
-    RclNode& node,
-    const GenericPubMap& pub_map,
-    const RobotParams& params,
-    const HopperState& hopper_state,
-    TraversalController& trav_controller) :
-    pub_map{pub_map},
-    params{params},
-    hopper_state{hopper_state},
-    traversal_controller{trav_controller},
-    mining_controller{node, pub_map, params, hopper_state}
-{
-}
-
-void AutoMiningController::initialize() { this->stage = Stage::INITIALIZATION; }
-
-bool AutoMiningController::isFinished()
-{
-    return this->stage == Stage::FINISHED;
-}
-
-void AutoMiningController::setCancelled() { this->stage = Stage::FINISHED; }
-
-void AutoMiningController::iterate(
-    const RobotMotorStatus& motor_status,
-    RobotMotorCommands& commands)
-{
-    switch (this->stage)
-    {
-        case Stage::INITIALIZATION:
-        {
-            this->stage = Stage::PLANNING;
-            [[fallthrough]];
-        }
-        case Stage::PLANNING:
-        {
-            if (false)  // *if not finished planning*
-            {
-                // call query service, wait for response, determine best option
-
-                break;  // break if more work is required
-            }
-
-            // init with planned destination
-            this->traversal_controller.initialize();
-            this->stage = Stage::TRAVERSING;
-            [[fallthrough]];
-        }
-        case Stage::TRAVERSING:
-        {
-            this->traversal_controller.iterate(motor_status, commands);
-            if (!this->traversal_controller.isFinished())
-            {
-                break;
-            }
-
-            // initialize with query result
-            this->mining_controller.initialize(0.f);
-            this->stage = Stage::MINING;
-            [[fallthrough]];
-        }
-        case Stage::MINING:
-        {
-            // this->mining_controller.setRemaining(); // <-- update remaining distance
-            this->mining_controller.iterate(motor_status, commands);
-            if (!this->mining_controller.isFinished())
-            {
-                break;
-            }
-
-            this->stage = Stage::FINISHED;
-            [[fallthrough]];
-        }
-        case Stage::FINISHED:
-        {
-        }
-    }
-}
