+++ lance/src/robot/controllers/auto_offload_controller.hpp
@@ -94,88 +94,3 @@ protected:
     TraversalController& traversal_controller;
     OffloadController offload_controller;
 };
-
-
-// --- Implementation ----------------------------------------------------------
-
-AutoOffloadController::AutoOffloadController(
-    RclNode& node,
-    const GenericPubMap& pub_map,
-    const RobotParams& params,
-    const HopperState& hopper_state,
-    TraversalController& trav_controller) :
-    pub_map{pub_map},
-    params{params},
-    hopper_state{hopper_state},
-    traversal_controller{trav_controller},
-    offload_controller{node, pub_map, params, hopper_state}
-{
-}
-
-void AutoOffloadController::initialize()
-{
-    this->stage = Stage::INITIALIZATION;
-}
-
-bool AutoOffloadController::isFinished()
-{
-    return this->stage == Stage::FINISHED;
-}
-
-void AutoOffloadController::setCancelled() { this->stage = Stage::FINISHED; }
-
-void AutoOffloadController::iterate(
-    const JoyState& joy,
-    const RobotMotorStatus& motor_status,
-    RobotMotorCommands& commands)
-{
-    switch (this->stage)
-    {
-        case Stage::INITIALIZATION:
-        {
-            this->stage = Stage::PLANNING;
-            [[fallthrough]];
-        }
-        case Stage::PLANNING:
-        {
-            if (false)  // *if not finished planning*
-            {
-                // planning algo here
-
-                break;  // break if more work is required
-            }
-
-            // init with planned destination
-            this->traversal_controller.initialize();
-            this->stage = Stage::TRAVERSING;
-            [[fallthrough]];
-        }
-        case Stage::TRAVERSING:
-        {
-            this->traversal_controller.iterate(motor_status, commands);
-            if (!this->traversal_controller.isFinished())
-            {
-                break;
-            }
-
-            // initialize with query result
-            this->offload_controller.initialize();
-            this->stage = Stage::OFFLOADING;
-            [[fallthrough]];
-        }
-        case Stage::OFFLOADING:
-        {
-            this->offload_controller.iterate(joy, motor_status, commands);
-            if (!this->offload_controller.isFinished())
-            {
-                break;
-            }
-
-            this->stage = Stage::FINISHED;
-            [[fallthrough]];
-        }
-        case Stage::FINISHED:
-        {
-        }
-    }
-}
