+++ lance/src/robot/controllers/localization_controller.hpp
@@ -79,48 +79,3 @@ protected:
 
     Stage stage{Stage::FINISHED};
 };
-
-
-// --- Implementation ----------------------------------------------------------
-
-LocalizationController::LocalizationController(
-    RclNode& node,
-    const GenericPubMap& pub_map,
-    const RobotParams& params) :
-    pub_map{pub_map},
-    params{params}
-{
-}
-
-void LocalizationController::initialize()
-{
-    this->stage = Stage::INITIALIZATION;
-}
-
-bool LocalizationController::isFinished()
-{
-    return this->stage == Stage::FINISHED;
-}
-
-void LocalizationController::setCancelled() {}
-
-void LocalizationController::iterate(
-    const RobotMotorStatus& motor_status,
-    RobotMotorCommands& commands)
-{
-    switch (this->stage)
-    {
-        case Stage::INITIALIZATION:
-        {
-        }
-        case Stage::SEARCHING:
-        {
-        }
-        case Stage::TARGETTING:
-        {
-        }
-        case Stage::FINISHED:
-        {
-        }
-    }
-}
