+++ lance/src/robot/controllers/offload_controller.hpp
@@ -69,6 +69,8 @@ public:
     bool isFinished();
     void setCancelled();
 
+    void setUseHopperModel(bool enabled);
+
     void iterate(
         const RobotMotorStatus& motor_status,
         RobotMotorCommands& commands);
@@ -112,107 +114,6 @@ protected:
 
     Stage stage{Stage::FINISHED};
     TraversalState traversal_state{};
+    double calculated_target_belt_pos{0.};
+    bool using_hopper_model{true};
 };
-
-
-// --- Implementation ----------------------------------------------------------
-
-OffloadController::OffloadController(
-    RclNode& node,
-    const GenericPubMap& pub_map,
-    const RobotParams& params,
-    const HopperState& hopper_state) :
-    pub_map{pub_map},
-    params{params},
-    hopper_state{hopper_state}
-{
-}
-
-void OffloadController::initialize(float traversal_dist_m)
-{
-    this->stage = Stage::INITIALIZATION;
-    this->traversal_state.init(traversal_dist_m);
-}
-
-bool OffloadController::isFinished() { return this->stage == Stage::FINISHED; }
-
-void OffloadController::setCancelled()
-{
-    this->stage = Stage::FINISHED;
-}
-
-void OffloadController::iterate(
-    const RobotMotorStatus& motor_status,
-    RobotMotorCommands& commands)
-{
-    this->iterate(nullptr, motor_status, commands);
-}
-
-void OffloadController::iterate(
-    const JoyState& joy,
-    const RobotMotorStatus& motor_status,
-    RobotMotorCommands& commands)
-{
-    this->iterate(&joy, motor_status, commands);
-}
-
-
-void OffloadController::TraversalState::init(float remaining_dist)
-{
-    this->remaining_dist = remaining_dist;
-    this->prev_odom = std::numeric_limits<float>::infinity();
-}
-void OffloadController::TraversalState::updateOdom(float odom)
-{
-    if (this->remaining_dist > 0.f && !std::isinf(this->prev_odom))
-    {
-        this->remaining_dist -= (this->prev_odom - odom);
-    }
-    this->prev_odom = odom;
-}
-bool OffloadController::TraversalState::hasRemaining()
-{
-    return this->remaining_dist > 0.f;
-}
-
-
-void OffloadController::iterate(
-    const JoyState* joy,
-    const RobotMotorStatus& motor_status,
-    RobotMotorCommands& commands)
-{
-    using namespace Bindings;
-
-    if(this->stage != Stage::INITIALIZATION && joy &&
-        AssistedOffloadToggleButton::wasPressed(*joy))
-    {
-        this->stage = Stage::LOWERING;
-    }
-
-    this->traversal_state.updateOdom(
-        static_cast<float>(track_motor_rps_to_ground_mps(
-            0.5 * (motor_status.track_left.position +
-                   motor_status.track_right.position))));
-
-    switch (this->stage)
-    {
-        case Stage::INITIALIZATION:
-        {
-        }
-        case Stage::BACKUP:
-        {
-        }
-        case Stage::RAISING:
-        {
-        }
-        case Stage::OFFLOADING:
-        {
-        }
-        case Stage::LOWERING:
-        {
-        }
-        case Stage::FINISHED:
-        {
-        }
-    }
-}
