+++ lance/src/robot/controllers/traversal_controller.hpp
@@ -78,45 +78,3 @@ protected:
 
     State state{State::FINISHED};
 };
-
-
-// --- Implementation ----------------------------------------------------------
-
-TraversalController::TraversalController(
-    RclNode& node,
-    const GenericPubMap& pub_map,
-    const RobotParams& params) :
-    pub_map{pub_map},
-    params{params}
-{
-}
-
-void TraversalController::initialize() { this->state = State::INITIALIZATION; }
-
-bool TraversalController::isFinished()
-{
-    return this->state == State::FINISHED;
-}
-
-void TraversalController::setCancelled() {}
-
-void TraversalController::iterate(
-    const RobotMotorStatus& motor_status,
-    RobotMotorCommands& commands)
-{
-    switch (this->state)
-    {
-        case State::INITIALIZATION:
-        {
-            //
-        }
-        case State::TRAVERSING:
-        {
-            //
-        }
-        case State::FINISHED:
-        {
-            //
-        }
-    }
-}
